feature(context-application) : move context application code to applications/context
