+++ src/collection.json
@@ -100,15 +100,15 @@
       "description": "Create a Nest Web Application.",
       "schema": "./lib/applications/web/schema.json"
     },
+    "context-application": {
+      "factory": "./lib/applications/context/factory#main",
+      "description": "Create a Nest Context Application.",
+      "schema": "./lib/applications/context/schema.json"
+    },
     "tcp-microservice-application": {
       "factory": "./lib/tcp-microservice-application/tcp-microservice-application.factory#main",
       "description": "Create a Nest TCP Microservice Application.",
       "schema": "./lib/tcp-microservice-application/schema.json"
-    },
-    "context-application": {
-      "factory": "./lib/context-application/context-application.factory#main",
-      "description": "Create a Nest Context Application.",
-      "schema": "./lib/context-application/schema.json"
     }
   }
 }
