+++ src/lib/applications/context/options.d.ts
@@ -23,12 +23,4 @@ export interface ContextApplicationOptions {
    * The used package manager.
    */
   packageManager?: 'npm' | 'yarn';
-  /**
-   * The Nest included production dependencies (comma separated values).
-   */
-  dependencies?: string;
-  /**
-   * The Nest included development dependencies (comma separated values).
-   */
-  devDependencies?: string;
 }
