--- src/lib/context-application/files/js/.prettierrc
@@ -1,4 +1,0 @@
-{
-  "singleQuote": true,
-  "trailingComma": "all"
-}
