--- src/lib/context-application/files/js/README.md
@@ -1,37 +1,0 @@
-# <%= name %>
-
-## Description
-
-<%= description %>
-
-## Installation
-
-```bash
-$ <%= packageManager %> install
-```
-
-## Running the app
-
-```bash
-# development
-$ <%= packageManager %> run start
-
-# watch mode
-$ <%= packageManager %> run start:dev
-
-# production mode
-<%= packageManager %> run start:prod
-```
-
-## Test
-
-```bash
-# unit tests
-$ <%= packageManager %> run test
-
-# e2e tests
-$ <%= packageManager %> run test:e2e
-
-# test coverage
-$ <%= packageManager %> run test:cov
-```
