--- src/lib/context-application/files/js/jsconfig.json
@@ -1,9 +1,0 @@
-{
-  "compilerOptions": {
-      "target": "ES6",
-      "experimentalDecorators": true
-  },
-  "exclude": [
-      "node_modules"
-  ]
-}
