--- src/lib/context-application/files/js/package.json
@@ -1,42 +1,0 @@
-{
-  "name": "<%= name %>",
-  "version": "<%= version %>",
-  "description": "<%= description %>",
-  "author": "<%= author %>",
-  "license": "MIT",
-  "scripts": {
-    "format": "prettier --write \"**/*.js\"",
-    "start": "babel-node index.js",
-    "start:dev": "nodemon",
-    "test": "jest",
-    "test:cov": "jest --coverage",
-    "test:e2e": "jest --config ./test/jest-e2e.json"
-  },
-  "dependencies": {
-    "@nestjs/common": "^6.2.4",
-    "@nestjs/core": "^6.2.4",
-    "reflect-metadata": "^0.1.13",
-    "rimraf": "^2.6.2",
-    "rxjs": "^6.5.2"
-  },
-  "devDependencies": {
-    "@nestjs/testing": "^6.0.0",
-    "@babel/core": "^7.4.3",
-    "@babel/node": "^7.2.2",
-    "@babel/plugin-proposal-decorators": "^7.4.0",
-    "@babel/preset-env": "^7.4.3",
-    "@babel/register": "^7.4.0",
-    "jest": "^22.4.3",
-    "nodemon": "^1.14.1",
-    "prettier": "^1.11.1"
-  },
-  "jest": {
-    "moduleFileExtensions": [
-      "js",
-      "json"
-    ],
-    "rootDir": "src",
-    "testRegex": ".spec.js$",
-    "coverageDirectory": "../coverage"
-  }
-}
