--- src/lib/context-application/files/ts/.prettierrc
@@ -1,4 +1,0 @@
-{
-  "singleQuote": true,
-  "trailingComma": "all"
-}
