--- src/lib/context-application/files/ts/nodemon.json
@@ -1,5 +1,0 @@
-{
-  "watch": ["dist"],
-  "ext": "js",
-  "exec": "node dist/main"
-}
