--- src/lib/context-application/files/ts/package.json
@@ -1,54 +1,0 @@
-{
-  "name": "<%= name %>",
-  "version": "<%= version %>",
-  "description": "<%= description %>",
-  "author": "<%= author %>",
-  "license": "MIT",
-  "scripts": {
-    "build": "tsc -p tsconfig.build.json",
-    "format": "prettier --write \"src/**/*.ts\"",
-    "start": "ts-node -r tsconfig-paths/register src/main.ts",
-    "start:dev": "concurrently --handle-input \"wait-on dist/main.js && nodemon\" \"tsc -w -p tsconfig.build.json\" ",
-    "start:debug": "nodemon --config nodemon-debug.json",
-    "prestart:prod": "rimraf dist && npm run build",
-    "start:prod": "node dist/main.js",
-    "lint": "tslint -p tsconfig.json -c tslint.json",
-    "test": "jest",
-    "test:watch": "jest --watch",
-    "test:cov": "jest --coverage",
-    "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
-    "test:e2e": "jest --config ./test/jest-e2e.json"
-  },
-  "dependencies": {
-    "@nestjs/common": "^6.2.4",
-    "@nestjs/core": "^6.2.4",
-    "reflect-metadata": "^0.1.13",
-    "rimraf": "^2.6.2",
-    "rxjs": "^6.5.2"
-  },
-  "devDependencies": {
-    "@nestjs/testing": "^6.2.4",
-    "@types/jest": "^24.0.13",
-    "@types/node": "^12.0.2",
-    "concurrently": "^4.1.0",
-    "jest": "^24.8.0",
-    "nodemon": "^1.19.0",
-    "prettier": "^1.17.1",
-    "ts-jest": "^24.0.2",
-    "ts-node": "^8.1.0",
-    "tsconfig-paths": "^3.8.0",
-    "tslint": "^5.16.0",
-    "typescript": "^3.4.5",
-    "wait-on": "^3.2.0"
-  },
-  "jest": {
-    "moduleFileExtensions": ["js", "json", "ts"],
-    "rootDir": "src",
-    "testRegex": ".spec.ts$",
-    "transform": {
-      "^.+\\.(t|j)s$": "ts-jest"
-    },
-    "coverageDirectory": "../coverage",
-    "testEnvironment": "node"
-  }
-}
